Added nCountThreshold to GetBudget
This change should be done to avoid finalizing budgets that don't pass the nCountThreshold threshold. Changing the > to >= will also eliminate the minimum requirement of 10 masternodes which is most likely an unintentional check. If there should be a minimum requirements of masternodes, I think it should be clarified and implemented through another parameter.